% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_density}
\alias{fit_density}
\title{Fit (multivariate) conditional density}
\usage{
fit_density(X, Y, input_data, nbins = getopt("nbins"),
  bin_estimator = getopt("bin_estimator"),
  verbose = getOption("densier.verbose"))
}
\arguments{
\item{X}{A vector containing the names of predictor variables to use for modeling.}

\item{Y}{A character string name of the column that represent the response variable(s) in the model.}

\item{input_data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{nbins}{...}

\item{bin_estimator}{....}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console.
Turn this on by default using \code{options(densier.verbose=TRUE)}.}
}
\value{
An R6 object containing the model fit(s).
}
\description{
Fit (multivariate) conditional density
}

