% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getopt}
\alias{getopt}
\title{Get Option Value for \code{condensier}}
\usage{
getopt(optname)
}
\arguments{
\item{optname}{The name of the consdensier option.}
}
\value{
An option value.
}
\description{
Get Option Value for \code{condensier}
}
\seealso{
\code{\link{condensier_options}}
}
