% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{predict_probability}
\alias{predict_probability}
\title{Predict probability (likelihood) for existing fit, given new data}
\usage{
predict_probability(model_fit, newdata)
}
\description{
Predict probability (likelihood) for existing fit, given new data
}
