% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_condensier_opts}
\alias{print_condensier_opts}
\title{Print Current Option Settings for \code{condensier}}
\usage{
print_condensier_opts()
}
\value{
Invisibly returns a list of \code{condensier} options.
}
\description{
Print Current Option Settings for \code{condensier}
}
\seealso{
\code{\link{condensier_options}}
}

