% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_densier_opts}
\alias{print_densier_opts}
\title{Print Current Option Settings for \code{densier}}
\usage{
print_densier_opts()
}
\value{
Invisibly returns a list of \code{densier} options.
}
\description{
Print Current Option Settings for \code{densier}
}
\seealso{
\code{\link{densier_options}}
}

