% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sample_value}
\alias{sample_value}
\title{Sample values from existing conditional density fit, given new data}
\usage{
sample_value(model_fit, newdata)
}
\description{
Sample values from existing conditional density fit, given new data
}

