% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3Class.R
\docType{class}
\name{sl3_wrapper_logisfitR6}
\alias{sl3_wrapper_logisfitR6}
\title{sl3_wrapper_logisfitR6}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sl3_wrapper_logisfitR6
}
\description{
sl3_wrapper_logisfitR6
}
\keyword{datasets}
